HMAN-81 refactored code
